deep-research-agent/pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "firecrawl-py>=4.8.0",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.2.1",
]
